feat: 이미지 fade 애니메이션 구현
